prospecting_engine/social_graph_v3/requirements.txt
python-dotenv==1.0.0
praw==7.7.1
requests==2.31.0
google-generativeai==0.3.2
phantombuster==0.3.0
